fix: change state initialization lifecycle
Request change for android.

Activity is calling save instance state on child
fragments before the view was attached
when the currentState was uninitialized which
leads to crash.

We are using currentState in onSaveInstanceState to keep
state in android.

Presenter's initializeState gives flexibility to init a state
depending on platform specific requirements.

Also fixed type, expose exceptions classes to give possibility to
handle exceptions from library.